refactor(manager): 466796 - PR changes - reorder service tests
